% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.coordinates.R
\name{ds.coordinates}
\alias{ds.coordinates}
\title{Wrapper for \code{coordinates()} function from sp package}
\usage{
ds.coordinates(x = NULL, coords = NULL, newobj = NULL,
  datasources = NULL)
}
\arguments{
\item{x}{name of an object on the server side of class "data.frame" to be
converted to a spatial object}

\item{coords}{spatial coordinates; column names (e.g. c("x","y")) specifying
which columns in object are the spatial coordinates.
If the coordinates are part of object, giving the reference does not
duplicate them, giving their value does duplicate them in the resulting structure.}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.coords'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
usually an object of class SpatialPointsDataFrame; if the coordinates
set cover the full set of variables in object, an object of class SpatialPoints
is returned
}
\description{
This function is a wrapper for the coordinates() function from the
sp package
}
\details{


See the \code{coordinates()} function from sp package for more details
}
\examples{
{

# Convert data frame D to a SpatialPointsDataFrame
# Data frame D has columns Lon and Lat which contain the coordinates

myvect <- c("Lon","Lat")
ds.coordinates('D',myvect)

}
}
\author{
Bishop, T.
}

