% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.coordsToLines.R
\name{ds.coordsToLines}
\alias{ds.coordsToLines}
\title{Function to convert groups of points to lines}
\usage{
ds.coordsToLines(coords = NULL, group = NULL, newobj = NULL,
  datasources = NULL)
}
\arguments{
\item{coords}{object of class SpatialPointsDataFrame to be converted
to a SpatialLinesDataFrame}

\item{group}{string indicating the name of the column in the data frame that defines
the grouping of the sets of points}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.lines'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
an object of class SpatialLines, each line having an ID taken from the column
specified by the group variable
}
\description{
This function converts a SpatialPointsDataFrame object into a
 SpatialLines object
}
\details{


The input object consists of sets of points grouped by an identifier. The
function takes each group of points and converts them into a line with an associated
identifier
}
\examples{
{

# Load log in data

data(GEOSPATIAL_logindata)

# login
# (by default the assigned dataset is a dataframe named 'D')
opals <- datashield.login(logins=GEOSPATIAL_logindata,assign=TRUE)

# Convert data frame D to a SpatialPointsDataFrame
# Data frame D has columns Lon and Lat which contain the coordinates

myvect <- c("Lon","Lat")
ds.coordinates('D',myvect, newobj='coords')

# Assign epsg coordinate system to a SpatialPointsDataFrame called mySPframe
# The coordinate system 4326 is the code for WGS84 (GPS)

ds.proj4string('coords',4326,'mySPframe')

# Transform epsg coordinate system for a SpatialPointsDataFrame called mySPframe
# The coordinate system 29902 is the code for Ireland

ds.spTransform('mySPframe',29902,'transformed')

# Transform groups of points in a SpatialPointsDataFrame called transformed into lines
# The points in the data frame are grouped by an idenifier in a column called id

ds.coordsToLines('transformed','person_id',newobj='my_lines')

}
}
\author{
Bishop, T.
}

