% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.gBuffer.R
\name{ds.gBuffer}
\alias{ds.gBuffer}
\title{Wrapper for \code{gBuffer()} function from sp and rgeos packages}
\usage{
ds.gBuffer(input = NULL, by_id = FALSE, ip_width = NULL, newobj = NULL,
  datasources = NULL)
}
\arguments{
\item{input}{name of a spatial object on the server side to which the buffer will be applied}

\item{by_id}{Logical determining if the function should be applied across subgeometries
(TRUE) or the entire object (FALSE)}

\item{ip_width}{Distance from original geometry to include in the new geometry.
Negative values are allowed. Either a numeric vector of length 1 when byid is
FALSE; if byid is TRUE: of length 1 replicated to the number of input geometries,
 or of length equal to the number of input geometries}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.buff'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
an object of class SpatialPolygonsDataFrame or SpatialPolygons,
depending on the class of input
}
\description{
This function is a wrapper for the \code{gBuffer()} function from the
sp and rgeos packages
}
\details{


See the \code{gBuffer()} function from rgeos package for more details
}
\examples{
{

# Load log in data

data(GEOSPATIAL_logindata)

# login
# (by default the assigned dataset is a dataframe named 'D')
opals <- datashield.login(logins=GEOSPATIAL_logindata,assign=TRUE)

# Convert data frame D to a SpatialPointsDataFrame
# Data frame D has columns Lon and Lat which contain the coordinates

myvect <- c("Lon","Lat")
ds.coordinates('D',myvect, newobj='coords')

# Assign epsg coordinate system to a SpatialPointsDataFrame called mySPframe
# The coordinate system 4326 is the code for WGS84 (GPS)

ds.proj4string('coords',4326,'mySPframe')

# Transform epsg coordinate system for a SpatialPointsDataFrame called mySPframe
# The coordinate system 29902 is the code for Ireland

ds.spTransform('mySPframe',29902,'transformed')

# Transform groups of points in a SpatialPointsDataFrame called transformed into lines
# The points in the data frame are grouped by an idenifier in a column called id

ds.coordsToLines('transformed','person_id',newobj='my_lines')

# Set a buffer of 150 around points contained in the bus_work object
# Buffer is created per id in the bus_work object

ds.gBuffer('my_lines',by_id=TRUE,ip_width=150,'my_buffer')

}
}
\author{
Bishop, T.
}

