% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.geometry.R
\name{ds.geometry}
\alias{ds.geometry}
\title{Wrapper for \code{geometry()} function from sp and rgeos packages}
\usage{
ds.geometry(input_x = NULL, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{input_x}{name of a spatial object on the server side to which the geometry will be applied}

\item{newobj.name}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.proj'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
an object of class SpatialPolygonsDataFrame or SpatialPolygons,
depending on the class of input
}
\description{
This function is a wrapper for the \code{geometry()} function from the
sp and rgeos packages
}
\details{
See the \code{geometry()} function from rgeos package for more details
}
\author{
Bishop, T.
}
