% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.geoSummary.R
\name{ds.geoSummary}
\alias{ds.geoSummary}
\title{#' @title Wrapper for \code{gBuffer()} function from sp and rgeos packages}
\usage{
ds.geoSummary(input = NULL, datasources = NULL)
}
\arguments{
\item{input}{name of a spatial object on the server side to which the buffer will be applied}

\item{datasources}{name of a}
}
\value{
an object of class SpatialPolygonsDataFrame or SpatialPolygons,
depending on the class of input
}
\description{
This function is a wrapper for the \code{gBuffer()} function from the
sp and rgeos packages
}
\details{
See the \code{gBuffer()} function from rgeos package for more details
}
\author{
Bishop, T.
}
