% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.lapply.R
\name{ds.lapply}
\alias{ds.lapply}
\title{Wrapper for \code{lapply()} function}
\usage{
ds.lapply(input = NULL, fun = NULL, newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{input}{name of a list}

\item{fun}{the function to apply as a character}

\item{newobj.name}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.list'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
nothing
}
\description{
This function is a wrapper for the \code{lapply()} function from the
sp and rgeos packages
}
\details{
See the \code{lapply()} function for more details
}
\author{
Bishop, T.
}
