% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.nothing.R
\name{ds.nothing}
\alias{ds.nothing}
\title{DOes nothing}
\usage{
ds.nothing(x, datasources = NULL)
}
\arguments{
\item{x}{name of the thing}

\item{datasources}{....}
}
\value{
a character vector of measured variables.
}
\description{
This function is similar to something
}
\examples{

}
