% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.over.R
\name{ds.over}
\alias{ds.over}
\title{Wrapper for \code{over()} function from sp and rgeos packages}
\usage{
ds.over(x = NULL, y = NULL, fn = NULL, returnList = FALSE,
  newobj = NULL, datasources = NULL)
}
\arguments{
\item{x}{geometry (locations) of the queries (spatial object)}

\item{y}{layer from which the geometries or attributes are queried (spatial object)}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.over'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
either an object of class SpatialPointsDataFrame or SpatialPoints,
depending on the class of x
}
\description{
This function is a wrapper for the \code{over()} function from the
sp and rgeos packages. Consistent spatial overlay for points, grids and polygons:
at the spatial locations of object x retrieves the indexes or attributes from spatial
object y
}
\details{


See the \code{over()} function from sp package for more details
}
\examples{
{

# Return the points of interest that fall within each region in a new object 'res'
# 'my_regions' consists of several regions defined in a SpatialPolygonDataFrame
# 'my_poi' consists of points of interest in a SpatialPointsDataFrame

ds.over(x='my_regions',y='my_poi',newobj='res')

}
}
\author{
Bishop, T.
}

