% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GEOSPATIAL_logindata.R
\docType{data}
\name{GEOSPATIAL_logindata}
\alias{GEOSPATIAL_logindata}
\title{Information required to login to opal servers}
\format{A data frame where the number of servers corresponds to the number of rows
\describe{
 \item{\code{server}}{a character, the formal name of the study}
 \item{\code{url}}{URL of the opal server}
 \item{\code{user}}{a character, a formal username or a
                    path to a valid ssl certificate, if required}
 \item{\code{password}}{a character, a formal password or a
                        path to a valid ssl key if required}
 \item{\code{table}}{a character, the path to the opal datasource that holds the data
                     to analyse}
}}
\usage{
data(GEOSPATIAL_logindata)
}
\description{
A table of with 5 columns: study name, URL, username, password and opal datasource.
}
\details{

}

