% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.gBuffer.R
\name{ds.gBuffer}
\alias{ds.gBuffer}
\title{#' @title Wrapper for \code{gBuffer()} function from sp and rgeos packages}
\usage{
ds.gBuffer(input = NULL, by_id = FALSE, ip_width = NULL,
  newobj.name = NULL, datasources = NULL)
}
\arguments{
\item{input}{name of a spatial object on the server side to which the buffer will be applied}

\item{by_id}{Logical determining if the function should be applied across subgeometries
(TRUE) or the entire object (FALSE)}

\item{ip_width}{Distance from original geometry to include in the new geometry.
Negative values are allowed. Either a numeric vector of length 1 when byid is
FALSE; if byid is TRUE: of length 1 replicated to the number of input geometries,
 or of length equal to the number of input geometries}

\item{newobj.name}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.buff'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
an object of class SpatialPolygonsDataFrame or SpatialPolygons,
depending on the class of input
}
\description{
This function is a wrapper for the \code{gBuffer()} function from the
sp and rgeos packages
}
\details{
See the \code{gBuffer()} function from rgeos package for more details
}
\author{
Bishop, T.
}
