% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.SpatialLinesDataFrame.R
\name{ds.SpatialLinesDataFrame}
\alias{ds.SpatialLinesDataFrame}
\title{Wrapper for \code{SpatialLinesDataFrame()} function from sp package}
\usage{
ds.SpatialLinesDataFrame(lines = NULL, data = NULL, newobj = NULL,
  datasources = NULL)
}
\arguments{
\item{lines}{name of a list of objects on the server side of class "SpatialLines" that
will be transformed in a SpatialLineDataFrame object}

\item{data}{name of an object on the server side of class data frame that contains
the data that you wish to to attach to the lines. (N.B. need a row corresponding to
each line object in the list)}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.df'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
either an object of class SpatialPointsDataFrame or SpatialPoints,
depending on the class of x
}
\description{
This function is a wrapper for the \code{SpatialLinesDataFrame()} function from the
sp package
}
\details{


See the \code{SpatialLinesDataFrame()} function from sp package for more details
}
\examples{
{


}
}
\author{
Bishop, T.
}

