% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ds.spTransform.R
\name{ds.spTransform}
\alias{ds.spTransform}
\title{Wrapper for \code{spTransform()} function from spand rgdal packages}
\usage{
ds.spTransform(x = NULL, projStr = NULL, newobj = NULL,
  datasources = NULL)
}
\arguments{
\item{x}{name of an object on the server side of class "SpatialPointsDataFrame"
or "SpatialPoints" for which the epsg coordinate system will be transformed}

\item{projStr}{a valid proj4 epsg coordinate system identifier e.g. 29902 for
Ireland. (N.B. need to add validation to check the id)}

\item{newobj}{a character, the name of the new object which will be created
If no name is specified the default name is the name of the original data frame
followed by the suffix '.trans'.}

\item{datasources}{a list of opal object(s) obtained after login in to opal servers;
these objects hold also the data assign to R, as \code{dataframe}, from opal datasources.}
}
\value{
either an object of class SpatialPointsDataFrame or SpatialPoints,
depending on the class of x
}
\description{
This function is a wrapper for the \code{spTransform()} function from the
sp and rgdal packages
}
\details{


See the \code{spTransform()} function from sp package for more details
}
\examples{
{

# Load log in data

data(GEOSPATIAL_logindata)

# login
# (by default the assigned dataset is a dataframe named 'D')
opals <- datashield.login(logins=GEOSPATIAL_logindata,assign=TRUE)

# Convert data frame D to a SpatialPointsDataFrame
# Data frame D has columns Lon and Lat which contain the coordinates

myvect <- c("Lon","Lat")
ds.coordinates('D',myvect, newobj='coords')

# Assign epsg coordinate system to a SpatialPointsDataFrame called mySPframe
# The coordinate system 4326 is the code for WGS84 (GPS)

ds.proj4string('coords',4326,'mySPframe')

# Transform epsg coordinate system for a SpatialPointsDataFrame called mySPframe
# The coordinate system 29902 is the code for Ireland

ds.spTransform('mySPframe',29902,'transformed')

}
}
\author{
Bishop, T.
}

